Add files via upload
A computer has three commonly used resources designated A, B and C. Up to three processes designated X, Y and Z run on the computer and each makes periodic use of two of the three resources.
 Process X acquires A, then B, uses both and then releases both.
 Process Y acquires B, then C, uses both and then releases both.
 Process Z acquires C, then A, uses both and then releases both.
Write a program and check the following condition in your code. All the resources and processes must be given by user as input.
a) If two of these processes are running simultaneously on the machine, can a deadlock occur? If so, show the deadlock scenario.
b) Show a scenario in which deadlock occurs if all three processes are running simultaneously on the machine.
c) Modify the code for acquiring resources so that deadlock cannot occur with three processes running